+++ packages/ts-components/src/components/job-title/assets/TooltipIcon.tsx
