+++ packages/ts-components/src/components/job-title/assets/close-icon.tsx
