+++ packages/ts-newskit/src/index.ts
@@ -2,7 +2,6 @@
 export { Breadcrumb } from './components/navigation/breadcrumb';
 
 // MISC
-export { JobTitle } from './components/misc/job-title/job-title';
 export {
   UpdateButtonWithDelay
 } from './components/misc/update-button/update-button-with-delay';
